backend2/requirements.txt
django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.4
openpyxl>=3.1
python-decouple>=3.8